Josephus Problem And KnapSack
Josephus problem (or Josephus permutation) is a theoretical problem
related to a certain counting-out game.
People are standing in a circle waiting to be executed. Counting begins
at a specified point in the circle and proceeds around the circle in a
specified direction. After a specified number of people are skipped, the
next person is executed. The procedure is repeated with the remaining
people, starting with the next person, going in the same direction and
skipping the same number of people, until only one person remains, and
is freed.
The problem — given the number of people, starting point, direction, and
number to be skipped — is to choose the position in the initial circle
to avoid execution.

////////////////////////////////////////////////////////////////////////////////////////////////

Knapsack Problem

The knapsack problem or rucksack problem is a problem in combinatorial
optimization: Given a set of items, each with a weight and a value,
determine the number of each item to include in a collection so that the
total weight is less than or equal to a given limit and the total value
is as large as possible.